# openapi.yaml
openapi: 3.1.0
info:
  title: AIDecepticon Control Plane API
  version: 0.1.0
  description: API for managing organization-isolated deception deployments, sensors, tokens, incidents, encrypted secrets, audit events, identity domains, and integrations. Operator endpoints use X-AIDecepticon-Organization to select an authorized organization.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8787/api/v1
security:
  - controlPlaneSession: []
  - controlPlaneBearer: []
paths:
...
  /summary:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: getSummary
      summary: Read the command-center summary
      responses:
        '200': { description: Summary metrics }
  /deployments:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listDeployments
      summary: List deception deployments
      responses:
        '200': { description: Deployment collection }
    post:
      operationId: createDeployment
      summary: Deploy a deception blueprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, blueprintId, environment, location]
              properties:
                name: { type: string }
                blueprintId: { type: string }
                environment: { type: string }
                location: { type: string }
                mode: { type: string, enum: [agentless, cloud-native, sensor, package] }
                customizations: { type: object }
      responses:
        '201': { description: Deployment accepted }
        '400': { description: Invalid input }
  /tokens:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listTokens
      summary: List deception tokens
      responses:
        '200': { description: Token collection }
    post:
      operationId: createToken
      summary: Generate a canary token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                type: { type: string, enum: [document, credential, connection, cloud-key, api-key] }
                destination: { type: string }
                metadata: { type: object }
      responses:
        '201': { description: Token generated }
...
  /incidents/{id}:
    patch:
      operationId: updateIncident
      summary: Update incident workflow state
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                assignee: { type: string }
      responses:
        '200': { description: Incident updated }
        '404': { description: Incident not found }
  /sensors:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listSensors
      summary: List projection sensors
      responses:
        '200': { description: Sensor collection }
...
  /secrets:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listEncryptedSecrets
      summary: List encrypted-secret metadata without ciphertext or plaintext
      responses:
        '200': { description: Secret metadata collection }
    post:
      operationId: createEncryptedSecret
      summary: Encrypt and store a provider credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value]
              properties:
                name: { type: string, maxLength: 100 }
                type: { type: string, enum: [integration, cloud, identity, response, api] }
                description: { type: string, maxLength: 300 }
                value: { type: string, format: password, maxLength: 4096 }
      responses:
        '201': { description: Secret encrypted; metadata returned without ciphertext }
        '400': { description: Invalid secret input }
  /secrets/{secretId}/rotate:
    post:
      operationId: rotateEncryptedSecret
      summary: Replace a secret with a newly encrypted value
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/SecretId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string, format: password, maxLength: 4096 }
      responses:
        '200': { description: Secret value rotated; metadata returned }
        '404': { description: Secret not found }
  /secrets/{secretId}/verify:
    post:
      operationId: verifyEncryptedSecret
      summary: Decrypt server-side and compare the secret integrity fingerprint
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/SecretId'
      responses:
        '200': { description: Verification result without plaintext }
        '404': { description: Secret not found }
  /audit-events:
    get:
      operationId: listAdministrativeAuditEvents
      summary: List append-only audit events and verify the HMAC chain
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200': { description: Audit events and chain verification result }
...
  /beacon/{tokenId}:
    get:
      security: []
      operationId: triggerBeacon
      summary: Trigger a deception token interaction
      parameters:
        - name: tokenId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Interaction recorded }
        '404': { description: Token not found }
components:
...
  securitySchemes:
    controlPlaneSession:
      type: apiKey
      in: cookie
      name: __Host-aidecepticon.sid
      description: HTTP-only server-side operator session established through OIDC, SAML, or API-key exchange.
    controlPlaneBearer:
      type: http
      scheme: bearer
      description: Optional control-plane API key configured by the administrator.
    sensorBearer:
      type: http
      scheme: bearer
      description: Unique sensor access token returned once during enrollment.
  parameters:
    OrganizationId:
      name: X-AIDecepticon-Organization
      in: header
      required: false
      description: Authorized organization to make active. The server selects the configured default when omitted.
      schema: { type: string, pattern: '^org-' }
...
    SecretId:
      name: secretId
      in: path
      required: true
      schema: { type: string, pattern: '^sec-' }